Rename filename of FormControlElementClerk.js to FormControlElementInspector.js
