+++ lib/dom/FormElementInspector.js
@@ -1,4 +1,4 @@
-import FormControlElementClerk from '../domClerks/FormControlElementClerk.js'
+import FormControlElementClerk from '../domClerks/FormControlElementInspector.js'
 
 import HashBuilder from '../domClerks/HashBuilder.js'
