+++ lib/domClerks/BaseLegacyFormElementClerk.js
@@ -1,4 +1,4 @@
-import FormControlElementClerk from './FormControlElementClerk.js'
+import FormControlElementClerk from './FormControlElementInspector.js'
 import VariablesValidator from '../validator/ValueHashValidator.js'
 
 /**
